ref(installer): host systemd probes and cryptenroll out of installer.py
* lib/systemd.py: service_started/service_state/accessibility_tools_in_use and the
  NTP/WKD wait are live-ISO probes, none of them touch the target. utils/ is off
  limits, general.py imports utils.env and the sibling contract would cycle
* disk/cryptenroll.py: tpm2/fido2 enrollment shared a device-collection block and
  the transient keyfile dance; DiskEncryption.encrypted_dev_paths() owns the first,
  a context manager the second
* scripts import accessibility_tools_in_use from lib.systemd, _resolve no longer
  needs the late import that dodged pyparted
* test_tpm2_pin drives enroll_tpm2() directly instead of Installer.__new__
